oauth2-demo/app/requirements.txt
fastapi
uvicorn[standard]
pyjwt